+++ js/line-size-check.js
@@ -1,0 +1,5 @@
+//Функция проверки минимальной и максимальной длины строки
+
+function commentLength (stroke, maxLength) {
+  return stroke.length <= maxLength;
+}
